# specs/001-build-me-an/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Food Court Ordering System API
  description: API for managing food court merchants, menus, orders, and buyer sessions
  version: 1.0.0
  contact:
    name: Ngantri Development Team
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://ngantri.vercel.app/api
    description: Production server
paths:
...
  /merchants/orders/{orderId}/status:
    put:
      summary: Update order status
      operationId: updateOrderStatus
      tags: [Orders]
      security:
        - sessionAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/order-status"
      responses:
        "200":
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    OrderItemResponse:
      type: object
      required: [id, menu_name, price, quantity, subtotal]
      properties:
        id:
          type: string
          format: uuid
        menu_id:
          type: string
          format: uuid
        menu_name:
          type: string
        menu_image_url:
          type: string
          format: uri
          nullable: true
        price:
          type: integer
          description: Price at time of order in Indonesian Rupiah (IDR)
        quantity:
          type: integer
        subtotal:
          type: integer
          description: Price × quantity in Indonesian Rupiah (IDR)
    OrderResponse:
      type: object
      required:
        [
          id,
          order_number,
          merchant_name,
          status,
          total_quantity,
          total_amount,
          items,
          created_at,
          updated_at,
        ]
      properties:
        id:
          type: string
          format: uuid
        buyer_session_id:
          type: string
          format: uuid
        merchant_id:
          type: string
          format: uuid
        merchant_name:
          type: string
        merchant_image_url:
          type: string
          format: uri
          nullable: true
        order_number:
          type: string
          pattern: '^\d{6}-\d{4}$'
        status:
          $ref: "#/components/schemas/order-status"
        total_quantity:
          type: integer
        total_amount:
          type: integer
          description: Total in cents
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponse"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...